file restructure and createInvoice edit
